Abstract out save/load-related code into a class
